requirements.txt
selenium==4.7.2
python-dotenv==0.21.0
slack-sdk==3.19.5
